# website/static/swagger.yaml
openapi: 3.0.3
info:
  title: Simple Virtual Diary API
  description: API for a virtual diary.
  version: 1.0.0
paths:
  /login:
    post:
      summary: User login
      tags:
        - user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
        '401':
          description: Incorrect password or email does not exist

  /logout:
    get:
      summary: User logout
      tags:
        - user
      responses:
        '200':
          description: Successful logout

  /sign-up:
    post:
      summary: User sign-up
      tags:
        - user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                firstName:
                  type: string
                password1:
                  type: string
                password2:
                  type: string
      responses:
        '200':
          description: Successful sign-up
        '400':
          description: Invalid input

  /user:
    get:
      summary: Get user information
      tags:
        - user
      responses:
        '200':
          description: Successful operation
        '404':
          description: User not found
    delete:
      summary: Delete user and associated data
      tags:
        - user
      responses:
        '200':
          description: Successful operation
        '404':
          description: User not found
    post:
      summary: Update user information
      tags:
        - user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input

  /delete-note:
    delete:
      summary: Delete note
      tags:
        - note
      responses:
        '200':
          description: Successful operation
        '404':
          description: Note not found

  /delete-diary/{diary_id}:
    delete:
      summary: Delete diary
      tags:
        - diary
      parameters:
        - name: diary_id
          in: path
          required: true
          description: ID of the diary to delete
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
        '404':
          description: Diary not found

  /diary-notes/edit/{note_id}:
    put:
      summary: Update note by ID
      tags:
        - note
      parameters:
        - name: note_id
          in: path
          required: true
          description: ID of the note to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                editedNote:
                  type: string
      responses:
        '200':
          description: Successful operation
        '404':
          description: Note not found

  /diary-notes/{diary_id}:
    get:
      summary: Get diary notes
      tags:
        - diary
      parameters:
        - name: diary_id
          in: path
          required: true
          description: ID of the diary
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
        '404':
          description: Diary not found

    post:
      summary: Create a new note for a diary
      tags:
        - note
      parameters:
        - name: diary_id
          in: path
          required: true
          description: ID of the diary
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                note:
                  type: string
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input

  /:
    post:
      summary: Create diary
      tags:
        - diary
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                diary:
                  type: string
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input

    get:
      summary: Get all diaries for a user
      tags:
        - diary
      responses:
        '200':
          description: Successful operation
        '404':
          description: User not found

components:
  schemas:
    Diary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        user_id:
          type: integer

    User:
      type: object
      properties:
        userId:
          type: integer
        firstName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          maxLength: 20
          pattern: "^[a-zA-Z0-9]+$"

    Note:
      type: object
      properties:
        id:
          type: integer
        data:
          type: string
        date:
          type: string
          format: date
        user_id:
          type: integer
          
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string